+++ src/HappyZu.CloudStore.Core/Trip/Ticket.cs
@@ -32,22 +32,22 @@ namespace HappyZu.CloudStore.Trip
         /// <summary>
         /// 市场价
         /// </summary>
-        public double MarketPrice { get; set; }
+        public decimal MarketPrice { get; set; }
 
         /// <summary>
         /// 商城价
         /// </summary>
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         /// <summary>
         /// 成本价
         /// </summary>
-        public double CostPrice { get; set; }
+        public decimal CostPrice { get; set; }
 
         /// <summary>
         /// 代理商价格
         /// </summary>
-        public double AgentPrice { get; set; }
+        public decimal AgentPrice { get; set; }
 
         /// <summary>
         /// 可以支付预付款
@@ -57,7 +57,7 @@ namespace HappyZu.CloudStore.Trip
         /// <summary>
         /// 预付款金额
         /// </summary>
-        public double FrontMoneyPrice { get; set; }
+        public decimal FrontMoneyPrice { get; set; }
 
         /// <summary>
         /// 可以使用积分
